% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryCatchShiny.R
\name{on_error}
\alias{on_error}
\title{on_error}
\usage{
on_error(e, stop_with_req = TRUE, title = "Error", size = "l", ...)
}
\arguments{
\item{e}{object of \code{simpleError} class}

\item{stop_with_req}{logical (default \code{TRUE}). If \code{TRUE}, then
stop shiny code execution with \link[shiny]{req}}

\item{title}{The title of the modal.}

\item{size}{The size (width) of the modal. One of \code{"xs"} for extra small, \code{"s"}
for small (default), \code{"m"} for medium, or \code{"l"} for large.}

\item{...}{optional arguments for \link[shinyalert]{shinyalert}}
}
\description{
Open modal window with error message.
}
\examples{
\dontrun{
  tryCatchShiny::run_example("01-shiny-app")
}
}
