% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryCatchShiny.R
\name{recover_on_error}
\alias{recover_on_error}
\title{recover_on_error}
\usage{
recover_on_error()
}
\description{
Function to be used in combination with \code{shiny.error} option:
\code{options(shiny.error = tryCatchShiny::recover_on_error)}
}
\examples{
\dontrun{
  options(shiny.error = tryCatchShiny::recover_on_error)

  or run example app:
}

}
