% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryCatchShiny.R
\name{tryCatchShiny}
\alias{tryCatchShiny}
\title{tryCatchShiny}
\usage{
tryCatchShiny(expr, stop_with_req = TRUE, title = "Error", size = "l", ...)
}
\arguments{
\item{...}{}
}
\description{
tryCatchShiny
}
