% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryCatchShiny.R
\name{tryCatchShiny}
\alias{tryCatchShiny}
\title{tryCatchShiny}
\usage{
tryCatchShiny(expr, stop_with_req = TRUE, title = "Error", size = "l", ...)
}
\arguments{
\item{expr}{expression}

\item{stop_with_req}{logical (default \code{TRUE}). If \code{TRUE}, then
stop shiny code execution with \link[shiny]{req}}

\item{title}{The title of the modal.}

\item{size}{The size (width) of the modal. One of \code{"xs"} for extra small, \code{"s"}
for small (default), \code{"m"} for medium, or \code{"l"} for large.}

\item{...}{
  Arguments passed on to \code{\link[=on_error]{on_error}}
  \describe{
    \item{\code{e}}{object of \code{simpleError} class}
  }}
}
\description{
Open modal box on error in Shiny application.
\strong{Requires \link[shinyalert]{useShinyalert}} - otherwise modal won't pop up.
}
\examples{
\dontrun{
  tryCatchShiny::run_example("01-shiny-app")
}
}
